cleaned up the rest of the examples separated from their component
