+++ src/components/SidePanel/examples/02.with-defaults.jsx
