+++ src/components/SidePanel/examples/06.without-resize.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { SidePanel } from '../../../src/index.js';
+import { SidePanel } from '../../../index.js';
 
 export default class extends React.Component {
 	constructor(...args) {
