+++ src/components/SingleSelect/examples/11.array-options.jsx
