+++ src/components/SingleSelect/examples/12.stateless.jsx
