+++ src/components/Underline/examples/02.with-regex-match.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Underline } from '../../../src/index.js';
+import { Underline } from '../../../index.js';
 
 export default class extends React.Component {
 	render() {
